Most common subtitle tags are now supported

+++ libse/SsaStyle.cs
@@ -19,7 +19,7 @@ namespace Nikse.SubtitleEdit.Core
         public Color Outline { get; set; }
         public Color Background { get; set; }
         public int ShadowWidth { get; set; }
-        public int OutlineWidth { get; set; }
+        public double OutlineWidth { get; set; }
         public string Alignment { get; set; }
         public int MarginLeft { get; set; }
         public int MarginRight { get; set; }
